# Conflicts:
#	crime-means-assessment/src/main/java/uk/gov/justice/laa/crime/meansassessment/controller/MeansAssessmentController.java
#	crime-means-assessment/src/main/java/uk/gov/justice/laa/crime/meansassessment/service/MeansAssessmentService.java
#	crime-means-assessment/src/test/java/uk/gov/justice/laa/crime/meansassessment/controller/MeansAssessmentControllerTest.java
#	crime-means-assessment/src/test/java/uk/gov/justice/laa/crime/meansassessment/service/MeansAssessmentServiceTest.java

+++ crime-means-assessment/src/main/java/uk/gov/justice/laa/crime/meansassessment/staticdata/repository/AssessmentCriteriaDetailRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import uk.gov.justice.laa.crime.meansassessment.staticdata.entity.AssessmentCriteriaDetailEntity;
 
 @Repository
-public interface AssessmentCriteriaDetailRepository extends CrudRepository<AssessmentCriteriaDetailEntity, Long> {
+public interface AssessmentCriteriaDetailRepository extends CrudRepository<AssessmentCriteriaDetailEntity, Integer> {
 
 }
